+++ SilkWay/app/src/main/java/com/example/silkway/data/model/DetailsListItem.kt
@@ -18,8 +18,8 @@ sealed class DetailsListItem {
         val currency: String,
         val name: String,
         val section: String,
-        val currentAmountRequests: Long,
-        val minAmountRequests: Long,
+        val currentAmountRequests: Int,
+        val minAmountRequests: Int,
     ): DetailsListItem()
 
     data class Title(
